Move /epinio/about to /:product/c/:cluster/about
